# api/gen/http/openapi.yaml
swagger: "2.0"
info:
    title: PgQueryNarrative API
    description: Data Storyteller AI - Convert PostgreSQL SQL queries to business narratives
    version: v1
host: localhost:8080
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /api/v1/queries/saved:
        get:
            tags:
                - queries
            summary: list_saved queries
            description: List saved queries
            operationId: queries#list_saved
            parameters:
                - name: tags
                  in: query
                  required: false
                  type: array
                  items:
                    type: string
                  collectionFormat: multi
                - name: limit
                  in: query
                  required: false
                  type: integer
                  default: 50
                  maximum: 100
                  minimum: 1
                - name: offset
                  in: query
                  required: false
                  type: integer
                  default: 0
                  minimum: 0
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/SavedQueryList'
                        required:
                            - items
                            - limit
                            - offset
            schemes:
                - http
        post:
            tags:
                - queries
            summary: save queries
            description: Save a query
            operationId: queries#save
            parameters:
                - name: SaveRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/SaveQueryPayload'
                    required:
                        - name
                        - sql
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/SavedQuery'
                        required:
                            - id
                            - name
                            - sql
                            - created_at
                            - updated_at
            schemes:
                - http
    /api/v1/queries/saved/{id}:
        get:
            tags:
                - queries
            summary: get_saved queries
            description: Get saved query by ID
            operationId: queries#get_saved
            parameters:
                - name: id
                  in: path
                  required: true
                  type: string
                  format: uuid
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/SavedQuery'
                        required:
                            - id
                            - name
                            - sql
                            - created_at
                            - updated_at
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/NotFoundError'
                        required:
                            - name
                            - message
            schemes:
                - http
        delete:
            tags:
                - queries
            summary: delete_saved queries
            description: Delete a saved query
            operationId: queries#delete_saved
            parameters:
                - name: id
                  in: path
                  required: true
                  type: string
                  format: uuid
            responses:
                "204":
                    description: No Content response.
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/NotFoundError'
                        required:
                            - name
                            - message
            schemes:
                - http
...
definitions:
...
    NotFoundError:
        title: NotFoundError
        type: object
        properties:
            code:
                type: string
                example: Blanditiis maiores expedita harum quia.
            message:
                type: string
                example: Nesciunt debitis ut molestiae optio.
            name:
                type: string
                example: Sit debitis optio.
        example:
            code: Temporibus ea voluptates ipsa rerum laudantium ab.
            message: Tenetur odio dolor.
            name: Voluptatem doloremque expedita itaque natus.
        required:
            - name
            - message
...
    SaveQueryPayload:
        title: SaveQueryPayload
        type: object
        properties:
            description:
                type: string
                example: tjn
                maxLength: 500
            name:
                type: string
                example: d
                minLength: 1
                maxLength: 200
            sql:
                type: string
                example: ln
                minLength: 1
                maxLength: 10000
            tags:
                type: array
                items:
                    type: string
                    example: Occaecati totam nesciunt illo quia dolorum ea.
                example:
                    - In aliquam voluptatem.
                    - Qui eius.
        example:
            description: of0
            name: 0k
            sql: v
            tags:
                - Neque ut ipsum deleniti at voluptatem.
                - Error modi dolorum non cum optio.
        required:
            - name
            - sql
    SavedQuery:
        title: SavedQuery
        type: object
        properties:
            created_at:
                type: string
                example: "1991-08-23T11:46:07Z"
                format: date-time
            description:
                type: string
                example: Beatae perferendis architecto impedit aliquid exercitationem ex.
            id:
                type: string
                example: 8f67c390-f522-4c36-8f51-aaa1afffeb08
                format: uuid
            name:
                type: string
                example: Eaque commodi expedita itaque repellendus aut.
            sql:
                type: string
                example: In aut velit quidem odio quisquam neque.
            tags:
                type: array
                items:
                    type: string
                    example: Eveniet voluptatem excepturi.
                example:
                    - Dicta temporibus fuga quas iusto.
                    - Enim est autem.
                    - Officiis vitae iste voluptate dignissimos dolor voluptas.
                    - Cupiditate perspiciatis voluptates reprehenderit consequuntur.
            updated_at:
                type: string
                example: "1987-08-05T22:00:10Z"
                format: date-time
        example:
            created_at: "1979-06-09T17:31:21Z"
            description: Explicabo culpa deleniti voluptate unde.
            id: 70ce8e74-360a-4cbf-aa0d-bda39a8045a6
            name: Et ipsam.
            sql: In animi nulla voluptatem magnam voluptas.
            tags:
                - Sunt provident vitae autem accusantium quia magnam.
                - Velit sunt beatae et voluptas.
                - Numquam aut id beatae nihil.
                - Rerum dolores eum commodi ut.
            updated_at: "1989-03-14T17:35:39Z"
        required:
            - id
            - name
            - sql
            - created_at
            - updated_at
    SavedQueryList:
        title: SavedQueryList
        type: object
        properties:
            items:
                type: array
                items:
                    $ref: '#/definitions/SavedQuery'
                example:
                    - created_at: "1985-10-06T14:39:29Z"
                      description: Adipisci dolorem itaque cum.
                      id: be2c12d4-8bda-42b2-8c5e-18379d18c8e4
                      name: Et rerum.
                      sql: Molestiae sunt voluptatem qui non.
                      tags:
                        - Sint quo autem impedit magnam.
                        - Expedita ratione amet facilis neque.
                      updated_at: "1982-10-26T06:43:47Z"
                    - created_at: "1985-10-06T14:39:29Z"
                      description: Adipisci dolorem itaque cum.
                      id: be2c12d4-8bda-42b2-8c5e-18379d18c8e4
                      name: Et rerum.
                      sql: Molestiae sunt voluptatem qui non.
                      tags:
                        - Sint quo autem impedit magnam.
                        - Expedita ratione amet facilis neque.
                      updated_at: "1982-10-26T06:43:47Z"
            limit:
                type: integer
                example: 1243341385
                format: int32
            offset:
                type: integer
                example: 595415687
                format: int32
        example:
            items:
                - created_at: "1985-10-06T14:39:29Z"
                  description: Adipisci dolorem itaque cum.
                  id: be2c12d4-8bda-42b2-8c5e-18379d18c8e4
                  name: Et rerum.
                  sql: Molestiae sunt voluptatem qui non.
                  tags:
                    - Sint quo autem impedit magnam.
                    - Expedita ratione amet facilis neque.
                  updated_at: "1982-10-26T06:43:47Z"
                - created_at: "1985-10-06T14:39:29Z"
                  description: Adipisci dolorem itaque cum.
                  id: be2c12d4-8bda-42b2-8c5e-18379d18c8e4
                  name: Et rerum.
                  sql: Molestiae sunt voluptatem qui non.
                  tags:
                    - Sint quo autem impedit magnam.
                    - Expedita ratione amet facilis neque.
                  updated_at: "1982-10-26T06:43:47Z"
                - created_at: "1985-10-06T14:39:29Z"
                  description: Adipisci dolorem itaque cum.
                  id: be2c12d4-8bda-42b2-8c5e-18379d18c8e4
                  name: Et rerum.
                  sql: Molestiae sunt voluptatem qui non.
                  tags:
                    - Sint quo autem impedit magnam.
                    - Expedita ratione amet facilis neque.
                  updated_at: "1982-10-26T06:43:47Z"
                - created_at: "1985-10-06T14:39:29Z"
                  description: Adipisci dolorem itaque cum.
                  id: be2c12d4-8bda-42b2-8c5e-18379d18c8e4
                  name: Et rerum.
                  sql: Molestiae sunt voluptatem qui non.
                  tags:
                    - Sint quo autem impedit magnam.
                    - Expedita ratione amet facilis neque.
                  updated_at: "1982-10-26T06:43:47Z"
            limit: 122729893
            offset: 1168647286
        required:
            - items
            - limit
            - offset
...