# api/gen/http/queries/server/encode_decode.go
// Code generated by goa v3.24.1, DO NOT EDIT.
//
// queries HTTP server encoders and decoders
//
// Command:
// $ goa gen github.com/pgquerynarrative/pgquerynarrative/api/design

package server

import (
	"context"
	"errors"
	"io"
	"net/http"
	"strconv"

	queries "github.com/pgquerynarrative/pgquerynarrative/api/gen/queries"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeRunResponse returns an encoder for responses returned by the queries
// run endpoint.
func EncodeRunResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*queries.RunQueryResult)
		enc := encoder(ctx, w)
		body := NewRunResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeRunRequest returns a decoder for requests sent to the queries run
// endpoint.
func DecodeRunRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*queries.RunQueryPayload, error) {
	return func(r *http.Request) (*queries.RunQueryPayload, error) {
		var (
			body RunRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if errors.Is(err, io.EOF) {
				return nil, goa.MissingPayloadError()
			}
			var gerr *goa.ServiceError
			if errors.As(err, &gerr) {
				return nil, gerr
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidateRunRequestBody(&body)
		if err != nil {
			return nil, err
		}
		payload := NewRunQueryPayload(&body)

		return payload, nil
	}
}

// EncodeRunError returns an encoder for errors returned by the run queries
// endpoint.
func EncodeRunError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "validation_error":
			var res *queries.ValidationError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewRunValidationErrorResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusBadRequest)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// EncodeListSavedResponse returns an encoder for responses returned by the
// queries list_saved endpoint.
func EncodeListSavedResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*queries.SavedQueryList)
		enc := encoder(ctx, w)
		body := NewListSavedResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeListSavedRequest returns a decoder for requests sent to the queries
// list_saved endpoint.
func DecodeListSavedRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*queries.ListSavedPayload, error) {
	return func(r *http.Request) (*queries.ListSavedPayload, error) {
		var (
			tags   []string
			limit  int32
			offset int32
			err    error
		)
		qp := r.URL.Query()
		tags = qp["tags"]
		{
			limitRaw := qp.Get("limit")
			if limitRaw == "" {
				limit = 50
			} else {
				v, err2 := strconv.ParseInt(limitRaw, 10, 32)
				if err2 != nil {
					err = goa.MergeErrors(err, goa.InvalidFieldTypeError("limit", limitRaw, "integer"))
				}
				limit = int32(v)
			}
		}
		if limit < 1 {
			err = goa.MergeErrors(err, goa.InvalidRangeError("limit", limit, 1, true))
		}
		if limit > 100 {
			err = goa.MergeErrors(err, goa.InvalidRangeError("limit", limit, 100, false))
		}
		{
			offsetRaw := qp.Get("offset")
			if offsetRaw != "" {
				v, err2 := strconv.ParseInt(offsetRaw, 10, 32)
				if err2 != nil {
					err = goa.MergeErrors(err, goa.InvalidFieldTypeError("offset", offsetRaw, "integer"))
				}
				offset = int32(v)
			}
		}
		if offset < 0 {
			err = goa.MergeErrors(err, goa.InvalidRangeError("offset", offset, 0, true))
		}
		if err != nil {
			return nil, err
		}
		payload := NewListSavedPayload(tags, limit, offset)

		return payload, nil
	}
}

// EncodeSaveResponse returns an encoder for responses returned by the queries
// save endpoint.
func EncodeSaveResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*queries.SavedQuery)
		enc := encoder(ctx, w)
		body := NewSaveResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeSaveRequest returns a decoder for requests sent to the queries save
// endpoint.
func DecodeSaveRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*queries.SaveQueryPayload, error) {
	return func(r *http.Request) (*queries.SaveQueryPayload, error) {
		var (
			body SaveRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if errors.Is(err, io.EOF) {
				return nil, goa.MissingPayloadError()
			}
			var gerr *goa.ServiceError
			if errors.As(err, &gerr) {
				return nil, gerr
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidateSaveRequestBody(&body)
		if err != nil {
			return nil, err
		}
		payload := NewSaveQueryPayload(&body)

		return payload, nil
	}
}

// EncodeGetSavedResponse returns an encoder for responses returned by the
// queries get_saved endpoint.
func EncodeGetSavedResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*queries.SavedQuery)
		enc := encoder(ctx, w)
		body := NewGetSavedResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeGetSavedRequest returns a decoder for requests sent to the queries
// get_saved endpoint.
func DecodeGetSavedRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*queries.GetSavedPayload, error) {
	return func(r *http.Request) (*queries.GetSavedPayload, error) {
		var (
			id  string
			err error

			params = mux.Vars(r)
		)
		id = params["id"]
		err = goa.MergeErrors(err, goa.ValidateFormat("id", id, goa.FormatUUID))
		if err != nil {
			return nil, err
		}
		payload := NewGetSavedPayload(id)

		return payload, nil
	}
}

// EncodeGetSavedError returns an encoder for errors returned by the get_saved
// queries endpoint.
func EncodeGetSavedError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "not_found":
			var res *queries.NotFoundError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewGetSavedNotFoundResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusNotFound)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// EncodeDeleteSavedResponse returns an encoder for responses returned by the
// queries delete_saved endpoint.
func EncodeDeleteSavedResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		w.WriteHeader(http.StatusNoContent)
		return nil
	}
}

// DecodeDeleteSavedRequest returns a decoder for requests sent to the queries
// delete_saved endpoint.
func DecodeDeleteSavedRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*queries.DeleteSavedPayload, error) {
	return func(r *http.Request) (*queries.DeleteSavedPayload, error) {
		var (
			id  string
			err error

			params = mux.Vars(r)
		)
		id = params["id"]
		err = goa.MergeErrors(err, goa.ValidateFormat("id", id, goa.FormatUUID))
		if err != nil {
			return nil, err
		}
		payload := NewDeleteSavedPayload(id)

		return payload, nil
	}
}

// EncodeDeleteSavedError returns an encoder for errors returned by the
// delete_saved queries endpoint.
func EncodeDeleteSavedError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "not_found":
			var res *queries.NotFoundError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewDeleteSavedNotFoundResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusNotFound)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// marshalQueriesColumnInfoToColumnInfoResponseBody builds a value of type
// *ColumnInfoResponseBody from a value of type *queries.ColumnInfo.
func marshalQueriesColumnInfoToColumnInfoResponseBody(v *queries.ColumnInfo) *ColumnInfoResponseBody {
	res := &ColumnInfoResponseBody{
		Name: v.Name,
		Type: v.Type,
	}

	return res
}

// marshalQueriesChartSuggestionToChartSuggestionResponseBody builds a value of
// type *ChartSuggestionResponseBody from a value of type
// *queries.ChartSuggestion.
func marshalQueriesChartSuggestionToChartSuggestionResponseBody(v *queries.ChartSuggestion) *ChartSuggestionResponseBody {
	if v == nil {
		return nil
	}
	res := &ChartSuggestionResponseBody{
		ChartType: v.ChartType,
		Label:     v.Label,
		Reason:    v.Reason,
	}

	return res
}

// marshalQueriesPeriodComparisonItemToPeriodComparisonItemResponseBody builds
// a value of type *PeriodComparisonItemResponseBody from a value of type
// *queries.PeriodComparisonItem.
func marshalQueriesPeriodComparisonItemToPeriodComparisonItemResponseBody(v *queries.PeriodComparisonItem) *PeriodComparisonItemResponseBody {
	if v == nil {
		return nil
	}
	res := &PeriodComparisonItemResponseBody{
		Measure:          v.Measure,
		Current:          v.Current,
		Previous:         v.Previous,
		Change:           v.Change,
		ChangePercentage: v.ChangePercentage,
		Trend:            v.Trend,
	}

	return res
}

// marshalQueriesSavedQueryToSavedQueryResponseBody builds a value of type
// *SavedQueryResponseBody from a value of type *queries.SavedQuery.
func marshalQueriesSavedQueryToSavedQueryResponseBody(v *queries.SavedQuery) *SavedQueryResponseBody {
	res := &SavedQueryResponseBody{
		ID:          v.ID,
		Name:        v.Name,
		SQL:         v.SQL,
		Description: v.Description,
		CreatedAt:   v.CreatedAt,
		UpdatedAt:   v.UpdatedAt,
	}
	if v.Tags != nil {
		res.Tags = make([]string, len(v.Tags))
		for i, val := range v.Tags {
			res.Tags[i] = val
		}
	}

	return res
}